Updated Connection Settings popup
The Kinect Machine is currently written to always be the server. As
such, there is no need to have the Start Client popup.
The IP popup now functions as the personal IP that we need to type in
for the Hololens. That is to say, it isn't neccessary to provide our own
IP address for the Vive clients to connect to us, but it is neccessary
for us to provide our own IP address for the Hololens to connect to us.
So, the IP box functions purely for the second purpose.
The Port box functions as the port for Vive clients. If a Vive client
would like to connect, they need to type in the same port that the
server typed in. This port is NOT the port that the Hololens will be
connecting with. Because we do not provide a way for the Hololens to
cange their port in-app, the Kinect will not either (we don't want to
have it fiddled with). However, since the Vive can change their port,
the server can change the port to allow the Vive to connect as well.
Also changed a handful of comments.